- Билдер внес в класс configuration
- удалил лишние классы configuration
- классы мапперы в отдельном паккадже
