--- src/main/java/com/github/_1c_syntax/mdclasses/mdosource/edt/Configuration.java
@@ -1,25 +1,0 @@
-package com.github._1c_syntax.mdclasses.mdosource.edt;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.github._1c_syntax.mdclasses.metadata.additional.CompatibilityMode;
-import com.github._1c_syntax.mdclasses.metadata.additional.ScriptVariant;
-import lombok.Getter;
-
-
-@Getter
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Configuration
-        extends MDObjectBase {
-
-    protected ScriptVariant scriptVariant = ScriptVariant.ENGLISH;
-    protected CompatibilityMode compatibilityMode;
-    protected CompatibilityMode configurationExtensionCompatibilityMode;
-
-    protected String defaultRunMode;
-    protected String defaultLanguage;
-    protected String dataLockControlMode;
-    protected String objectAutonumerationMode;
-    protected String modalityUseMode;
-    protected String synchronousPlatformExtensionAndAddInCallUseMode;
-
-}
