+++ src/test/java/com/github/_1c_syntax/mdclasses/ConfigurationEDTTest.java
@@ -1,11 +1,10 @@
 package com.github._1c_syntax.mdclasses;
 
-import com.github._1c_syntax.mdclasses.metadata.ConfigurationBuilder;
+import com.github._1c_syntax.mdclasses.metadata.Configuration;
 import com.github._1c_syntax.mdclasses.metadata.additional.CompatibilityMode;
 import com.github._1c_syntax.mdclasses.metadata.additional.ConfigurationSource;
 import com.github._1c_syntax.mdclasses.metadata.additional.ModuleType;
 import com.github._1c_syntax.mdclasses.metadata.additional.ScriptVariant;
-import com.github._1c_syntax.mdclasses.metadata.configurations.AbstractConfiguration;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
@@ -15,54 +14,51 @@ import java.nio.file.Paths;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class ConfigurationEDTTest {
-  @Test
-  void testBuilder() {
+    @Test
+    void testBuilder() {
 
-    File srcPath = new File("src/test/resources/metadata/edt");
-    ConfigurationBuilder configurationBuilder = new ConfigurationBuilder(srcPath.toPath());
-    AbstractConfiguration configuration = configurationBuilder.build();
+        File srcPath = new File("src/test/resources/metadata/edt");
+        Configuration configuration = Configuration.newBuilder(srcPath.toPath()).build();
 
-    assertThat(configuration.getConfigurationSource()).isEqualTo(ConfigurationSource.EDT);
-    assertThat(configuration.getScriptVariant() == ScriptVariant.RUSSIAN).isTrue();
-    assertThat(CompatibilityMode.compareTo(configuration.getCompatibilityMode(), new CompatibilityMode(3, 10))).isEqualTo(0);
-    assertThat(configuration.getModulesByType().size() > 0).isTrue();
+        assertThat(configuration.getConfigurationSource()).isEqualTo(ConfigurationSource.EDT);
+        assertThat(configuration.getScriptVariant() == ScriptVariant.RUSSIAN).isTrue();
+        assertThat(CompatibilityMode.compareTo(configuration.getCompatibilityMode(), new CompatibilityMode(3, 10))).isEqualTo(0);
+        assertThat(configuration.getModulesByType().size() > 0).isTrue();
 
-    File file = new File("src/test/resources/metadata/edt/src/Constants/Константа1/ManagerModule.bsl");
-    assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.ManagerModule);
+        File file = new File("src/test/resources/metadata/edt/src/Constants/Константа1/ManagerModule.bsl");
+        assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.ManagerModule);
 
-    file = new File("src/test/resources/metadata/edt/src/CommonModules/ПростойОбщийМодуль/Module.bsl");
-    assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.CommonModule);
+        file = new File("src/test/resources/metadata/edt/src/CommonModules/ПростойОбщийМодуль/Module.bsl");
+        assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.CommonModule);
 
-  }
+    }
 
-  @Test
-  void testErrorBuild() {
+    @Test
+    void testErrorBuild() {
 
-    Path srcPath = Paths.get("src/test/resources/metadata");
-    ConfigurationBuilder configurationBuilder = new ConfigurationBuilder(srcPath);
-    AbstractConfiguration configuration = configurationBuilder.build();
+        Path srcPath = Paths.get("src/test/resources/metadata");
+        Configuration configuration = Configuration.newBuilder(srcPath).build();
 
-    assertThat(configuration).isNotNull();
+        assertThat(configuration).isNotNull();
 
-    File file = new File("src/test/resources/metadata/Module.os");
-    assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.Unknown);
-  }
+        File file = new File("src/test/resources/metadata/Module.os");
+        assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.Unknown);
+    }
 
-  @Test
-  void testBuilderEn() {
+    @Test
+    void testBuilderEn() {
 
-    File srcPath = new File("src/test/resources/metadata/edt_en");
-    ConfigurationBuilder configurationBuilder = new ConfigurationBuilder(srcPath.toPath());
-    AbstractConfiguration configuration = configurationBuilder.build();
+        File srcPath = new File("src/test/resources/metadata/edt_en");
+        Configuration configuration = Configuration.newBuilder(srcPath.toPath()).build();
 
-    assertThat(configuration.getConfigurationSource()).isEqualTo(ConfigurationSource.EDT);
-    assertThat(configuration.getScriptVariant() == ScriptVariant.ENGLISH).isTrue();
-    assertThat(CompatibilityMode.compareTo(configuration.getCompatibilityMode(), new CompatibilityMode(3, 14))).isEqualTo(0);
-    assertThat(configuration.getModulesByType().size() > 0).isTrue();
+        assertThat(configuration.getConfigurationSource()).isEqualTo(ConfigurationSource.EDT);
+        assertThat(configuration.getScriptVariant() == ScriptVariant.ENGLISH).isTrue();
+        assertThat(CompatibilityMode.compareTo(configuration.getCompatibilityMode(), new CompatibilityMode(3, 14))).isEqualTo(0);
+        assertThat(configuration.getModulesByType().size() > 0).isTrue();
 
-    File file = new File("src/test/resources/metadata/edt_en/src/CommonModules/CommonModule/Module.bsl");
-    assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.CommonModule);
+        File file = new File("src/test/resources/metadata/edt_en/src/CommonModules/CommonModule/Module.bsl");
+        assertThat(configuration.getModuleType(file.toURI())).isEqualTo(ModuleType.CommonModule);
 
-  }
+    }
 
 }
