+++ src/test/java/com/github/_1c_syntax/mdclasses/MDOSourceOriginTest.java
@@ -2,7 +2,7 @@ package com.github._1c_syntax.mdclasses;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
-import com.github._1c_syntax.mdclasses.mdosource.original.MetaDataObject;
+import com.github._1c_syntax.mdclasses.mdo.MetaDataObject;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
@@ -12,27 +12,25 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class MDOSourceOriginTest {
 
-  @Test
-  void testLoadConfiguration() {
+    @Test
+    void testLoadConfiguration() {
 
-    MetaDataObject MDObject = null;
-    String basePath = "src/test/resources/metadata/original";
-    File XML = new File(basePath, "Configuration.xml");
+        MetaDataObject MDObject = null;
+        String basePath = "src/test/resources/metadata/original";
+        File XML = new File(basePath, "Configuration.xml");
 
-    MetaDataObject mdObject = null;
+        XmlMapper xmlMapper = new XmlMapper();
+        xmlMapper.configure(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY, true);
 
-    XmlMapper xmlMapper = new XmlMapper();
-    xmlMapper.configure(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY, true);
+        try {
+            MDObject = xmlMapper.readValue(XML, MetaDataObject.class);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
 
-    try {
-      MDObject = xmlMapper.readValue(XML, MetaDataObject.class);
-    } catch (IOException e) {
-      e.printStackTrace();
+        assertThat(MDObject).isNotNull();
+        assertThat(MDObject.getConfiguration()).isNotNull();
+
     }
-
-    assertThat(MDObject).isNotNull();
-    assertThat(MDObject.getConfiguration()).isNotNull();
-
-  }
 
 }
